// src/api/generated/models/simpleStoryTagListItemResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Manyak Server API
 * Manyak backend API documentation
 * OpenAPI spec version: v1
 */
import type { SimpleStoryTagListItemResponseCategory } from './simpleStoryTagListItemResponseCategory';

/**
 * 간편 제작 태그 목록 항목
 */
export interface SimpleStoryTagListItemResponse {
  /** 태그 ID */
  tagId?: number;
  /** 태그 이름 */
  name?: string;
  /** 태그 분류 */
  category?: SimpleStoryTagListItemResponseCategory;
}